{{cookiecutter.github_repository}}/requirements/common.txt
# Core Stuff
# -------------------------------------
Django==1.10.5
django-flat-responsive==1.3.0

# WSGI HTTP Server
# -------------------------------------
uWSGI==2.0.14

# Configuration
# -------------------------------------
argon2-cffi==16.3.0
django-environ==0.4.1
django-sites==0.9
python-dotenv==0.6.3

# Staticfiles
# -------------------------------------
whitenoise==3.3.0

# Extensions
# -------------------------------------
pytz==2016.10

# Models
# -------------------------------------
psycopg2==2.6.2

Pillow==4.0.0
django-versatileimagefield==1.6.3
django-uuid-upload-path==1.0.0

# REST APIs
# -------------------------------------
djangorestframework==3.5.4

# LOGGING
# -------------------------------------
django-log-request-id==1.3.2

{% if cookiecutter.use_sentry_for_error_reporting == "y" -%}
# Raven is the Sentry client
# --------------------------
raven==6.0.0
{%- endif %}
{%- if cookiecutter.add_webpack.lower() == 'y' %}

# WEBPACK
# -------------------------------------
django-webpack-loader==0.3.3
{%- endif %}
